+++ src/prob/test.jl
@@ -45,6 +45,117 @@ function post_cl_opf(pm::GenericPowerModel)
 end
 
 
+"opf with unit commitment, tests constraint_current_limit"
+function run_uc_opf(file, model_constructor, solver; kwargs...)
+    return run_generic_model(file, model_constructor, solver, post_uc_opf; solution_builder = get_uc_solution, kwargs...)
+end
+
+""
+function post_uc_opf(pm::GenericPowerModel)
+    variable_voltage(pm)
+
+    variable_generation_indicator(pm)
+    variable_generation_on_off(pm)
+
+    variable_branch_flow(pm)
+    variable_dcline_flow(pm)
+
+    objective_min_fuel_cost(pm)
+
+    constraint_voltage(pm)
+
+    for i in ids(pm, :ref_buses)
+        constraint_theta_ref(pm, i)
+    end
+
+    for i in ids(pm, :gen)
+        constraint_generation_on_off(pm, i)
+    end
+
+    for i in ids(pm, :bus)
+        constraint_kcl_shunt(pm, i)
+    end
+
+    for i in ids(pm, :branch)
+        constraint_ohms_yt_from(pm, i)
+        constraint_ohms_yt_to(pm, i)
+
+        constraint_voltage_angle_difference(pm, i)
+
+        constraint_thermal_limit_from(pm, i)
+        constraint_thermal_limit_to(pm, i)
+    end
+
+    for i in ids(pm, :dcline)
+        constraint_dcline(pm, i)
+    end
+end
+
+
+""
+function run_uc_mc_opf(file, model_constructor, solver; kwargs...)
+    return run_generic_model(file, model_constructor, solver, post_uc_mc_opf; solution_builder = get_uc_solution, multiconductor=true, kwargs...)
+end
+
+""
+function post_uc_mc_opf(pm::GenericPowerModel)
+    variable_generation_indicator(pm)
+
+    for c in conductor_ids(pm)
+        variable_voltage(pm, cnd=c)
+        variable_voltage(pm, cnd=c)
+
+        variable_generation_on_off(pm, cnd=c)
+
+        variable_branch_flow(pm, cnd=c)
+        variable_dcline_flow(pm, cnd=c)
+
+        constraint_voltage(pm, cnd=c)
+
+        for i in ids(pm, :ref_buses)
+            constraint_theta_ref(pm, i, cnd=c)
+        end
+
+        for i in ids(pm, :gen)
+            constraint_generation_on_off(pm, i, cnd=c)
+        end
+
+        for i in ids(pm, :bus)
+            constraint_kcl_shunt(pm, i, cnd=c)
+        end
+
+        for i in ids(pm, :branch)
+            constraint_ohms_yt_from(pm, i, cnd=c)
+            constraint_ohms_yt_to(pm, i, cnd=c)
+
+            constraint_voltage_angle_difference(pm, i, cnd=c)
+
+            constraint_thermal_limit_from(pm, i, cnd=c)
+            constraint_thermal_limit_to(pm, i, cnd=c)
+        end
+
+        for i in ids(pm, :dcline)
+            constraint_dcline(pm, i, cnd=c)
+        end
+    end
+
+    objective_min_fuel_cost(pm)
+end
+
+""
+function get_uc_solution(pm::GenericPowerModel, sol::Dict{String,<:Any})
+    add_bus_voltage_setpoint(sol, pm)
+    add_generator_power_setpoint(sol, pm)
+    add_generator_status_setpoint(sol, pm)
+    add_storage_setpoint(sol, pm)
+    add_branch_flow_setpoint(sol, pm)
+    add_dcline_flow_setpoint(sol, pm)
+
+    add_kcl_duals(sol, pm)
+    add_sm_duals(sol, pm) # Adds the duals of the transmission lines' thermal limits.
+end
+
+
 ""
 function run_mn_opb(file, model_constructor, solver; kwargs...)
     return run_generic_model(file, model_constructor, solver, post_mn_opb; multinetwork=true, kwargs...)